+++ depthmapX/views/viewhelpers.cpp
@@ -7,8 +7,8 @@
 
 namespace ViewHelpers {
     Point2f calculateCenter(const QPointF &point, const QPointF &oldCentre, double factor) {
-        int diffX = oldCentre.x() - point.x();
-        int diffY = oldCentre.y() - point.y();
+        int diffX = static_cast<int>(oldCentre.x() - point.x());
+        int diffY = static_cast<int>(oldCentre.y() - point.y());
         return Point2f(point.x() + double(diffX) * factor, point.y() + double(diffY) * factor);
     }
